refactor(ui): remove components kept alive only by barrels or tests
Deletes ScopeInfoLine, GitFileListItem, GitFileTreeItem (its node type moves
to GitFileTreeNode.ts), MenuPanel and its ListPanelTabPillRow, ChartTooltip,
MultiSelectFooter and the shared layouts Section, and strips the unused
component bindings QuotaStatusBar, TaskUpdateCard, NestedActivityList's
default export, WorkItemDetailHeader, OverviewContent, the Dot* counters,
InlineOptionPill, PropertyFieldEditable's Dropdown and TextEditOption,
PlaygroundChatExtrasSection and ChatPreviewShell from files whose other
exports stay live.
