+++ src/components/ListPanel/index.ts
@@ -5,7 +5,6 @@
  */
 export { LIST_PANEL_SECTIONS, getListItemClasses } from "./tokens";
 
-export { MenuPanel } from "./MenuPanel";
 export { default as ListPanelItem } from "./ListPanelItem";
 
 export { ListPanelSkeletonRows } from "./ListPanelSkeletonRows";
