+++ src/components/PropertyField/PropertyFieldEditable.tsx
@@ -168,7 +168,7 @@ export const FieldRow: React.FC<FieldRowProps> = ({
 };
 
 // ============================================
-// Dropdown - Container for options (relative positioning)
+// Dropdown alignment helpers
 // ============================================
 
 export type DropdownWidthMode = "match-parent" | "menu";
@@ -215,43 +215,6 @@ function useResolvedDropdownAlign(align: DropdownAlign) {
     isPositioned: align !== "auto" || isAutoPositioned,
   };
 }
-
-export interface DropdownProps {
-  children: React.ReactNode;
-  className?: string;
-  align?: DropdownAlign;
-  widthMode?: DropdownWidthMode;
-}
-
-export const Dropdown: React.FC<DropdownProps> = ({
-  children,
-  className = "",
-  align = "left",
-  widthMode = "match-parent",
-}) => {
-  const { dropdownRef, resolvedAlign, isPositioned } =
-    useResolvedDropdownAlign(align);
-  const positionClass =
-    widthMode === "menu"
-      ? resolvedAlign === "right"
-        ? "right-0"
-        : "left-0"
-      : resolvedAlign === "right"
-        ? "right-2"
-        : "left-2 right-2";
-  const widthClass = widthMode === "menu" ? DROPDOWN_WIDTHS.wideMenuClass : "";
-
-  return (
-    <div
-      ref={dropdownRef}
-      data-property-dropdown
-      className={`absolute ${positionClass} top-full mt-1 flex flex-col ${widthClass} ${DROPDOWN_CLASSES.panelAnimated} ${className}`}
-      style={getPositionedOverlayVisibilityStyle(isPositioned)}
-    >
-      {children}
-    </div>
-  );
-};
 
 // ============================================
 // SearchableDropdown - Dropdown with search input (relative positioning)
@@ -443,77 +406,3 @@ export const Option: React.FC<OptionProps> = ({
     )}
   </button>
 );
-
-// ============================================
-// TextEditOption - Text area option in dropdown
-// ============================================
-
-export interface TextEditOptionProps {
-  value: string;
-  onChange: (value: string) => void;
-  onSubmit?: () => void;
-  onCancel?: () => void;
-  placeholder?: string;
-  rows?: number;
-}
-
-export const TextEditOption: React.FC<TextEditOptionProps> = ({
-  value,
-  onChange,
-  onSubmit,
-  onCancel,
-  placeholder = "Enter custom text...",
-  rows = 3,
-}) => {
-  const handleKeyDown = (event: React.KeyboardEvent<HTMLTextAreaElement>) => {
-    if (event.key === "Enter" && (event.metaKey || event.ctrlKey)) {
-      event.preventDefault();
-      onSubmit?.();
-    }
-    if (event.key === "Escape") {
-      event.preventDefault();
-      onCancel?.();
-    }
-    // Prevent dropdown from closing when typing
-    event.stopPropagation();
-  };
-
-  return (
-    <div className="px-2.5 py-2">
-      <textarea
-        value={value}
-        onChange={(event) => onChange(event.target.value)}
-        onKeyDown={handleKeyDown}
-        onClick={(event) => event.stopPropagation()}
-        placeholder={placeholder}
-        rows={rows}
-        className="w-full resize-none rounded-md border border-border-2 bg-bg-1 px-2 py-1.5 text-xs text-text-1 transition-colors outline-none placeholder:text-text-3 focus:border-primary-6"
-      />
-      <div className="mt-1.5 flex items-center justify-between gap-2">
-        <div className="flex gap-1">
-          <button
-            onClick={(event) => {
-              event.stopPropagation();
-              onSubmit?.();
-            }}
-            className="rounded bg-primary-6 px-2 py-0.5 text-[11px] text-white transition-colors hover:bg-primary-5"
-          >
-            Save
-          </button>
-          <button
-            onClick={(event) => {
-              event.stopPropagation();
-              onCancel?.();
-            }}
-            className="rounded bg-fill-2 px-2 py-0.5 text-[11px] text-text-2 transition-colors hover:bg-fill-2"
-          >
-            Cancel
-          </button>
-        </div>
-        <div className="text-[11px] text-text-3">
-          {navigator.platform.includes("Mac") ? "⌘" : "Ctrl"}+Enter
-        </div>
-      </div>
-    </div>
-  );
-};
