+++ src/engines/ChatPanel/blocks/ToolCallBlock/cards/index.ts
@@ -5,6 +5,6 @@ export { default as FileCard } from "./FileCard";
 export { default as SessionLinkCard } from "./SessionLinkCard";
 
 export { default as ProjectCard } from "./ProjectCard";
-export { TaskListCard, default as TaskUpdateCard } from "./TaskUpdateCard";
+export { TaskListCard } from "./TaskUpdateCard";
 export { default as WebsiteCard } from "./WebsiteCard";
 export { default as WorkItemCard } from "./WorkItemCard";
