+++ src/features/SessionCreator/components/index.ts
@@ -9,5 +9,3 @@ export { default as ControlButtons } from "./ControlButtons";
 export { default as EditorArea } from "./EditorArea";
 export { default as LaunchButton } from "./LaunchButton";
 export { default as SessionInfoLine } from "./SessionInfoLine";
-
-export { default as ScopeInfoLine } from "./ScopeInfoLine";
