+++ src/modules/MainApp/Integrations/DevTools/playground/panels/index.ts
@@ -1,4 +1,4 @@
-export { ChatPreviewShell, PlaygroundPreviewShell } from "./ChatPreviewShell";
+export { PlaygroundPreviewShell } from "./PlaygroundPreviewShell";
 export { PlaygroundChatPanel } from "./PlaygroundChatPanel";
 export * from "./PlaygroundLayout";
 export * from "./TokenOverridePanel";
