+++ src/modules/WorkStation/CodeEditor/Panels/EditorPrimarySidebar/content/SourceControlContent/hooks/useFileTreeHandling.ts
@@ -10,7 +10,7 @@ import { useCallback, useMemo, useState } from "react";
 
 import type { GitFile } from "@src/types/git/types";
 
-import type { GitFileTreeNode } from "../components/GitFileTreeItem";
+import type { GitFileTreeNode } from "../components/GitFileTreeNode";
 import { buildVSCodeStyleTree, flattenGitFileTree } from "../utils/treeUtils";
 
 export interface UseFileTreeHandlingOptions {
