+++ src/modules/WorkStation/CodeEditor/Panels/EditorPrimarySidebar/content/SourceControlContent/utils/virtualizedTreeUtils.ts
@@ -6,7 +6,7 @@
  */
 import type { GitFile } from "@src/types/git/types";
 
-import type { GitFileTreeNode } from "../components/GitFileTreeItem";
+import type { GitFileTreeNode } from "../components/GitFileTreeNode";
 import { buildVSCodeStyleTree, flattenGitFileTree } from "./treeUtils";
 
 // ============================================
