+++ src/modules/shared/layouts/blocks/ListPanelScrollArea.tsx
@@ -2,9 +2,9 @@
  * ListPanelScrollArea — Reusable scroll container for left list panels.
  *
  * Wraps list content with standard scroll behavior and configurable top padding.
- * Place below ListPanelTabPillRow or a panel search control.
+ * Place below a panel search control.
  *
- * Used by: MenuPanel, InboxListPanel, and similar panels.
+ * Used by: InboxListPanel and similar panels.
  */
 import React from "react";
