+++ src/modules/shared/layouts/index.ts
@@ -17,4 +17,3 @@ export { default as SplitViewLayout } from "./SplitViewLayout";
 export { default as SplitListFullscreenButton } from "./SplitListFullscreenButton";
 export { default as SplitListHeader } from "./SplitListHeader";
 export { default as InboxListDetailLayout } from "./InboxListDetailLayout";
-export { default as Section } from "./Section";
